Рефакторинг api продукта

+++ backend/src/services/productService.js
@@ -14,36 +14,38 @@ class ProductService {
     });
   }
 
-  async getAll() {
-    return await Product.findAll();
+  async getAll(userId) {
+    return await Product.findAll({ where: { userId } });
   }
 
-  async getOne(id) {
-    const product = await Product.findByPk(id);
+  async getOne(userId, productId) {
+    const product = await Product.findOne({ where: { userId, id: productId } });
     if (!product) {
       throw new ApiError(404, "Продукт с указанным id не найден");
     }
     return product;
   }
 
-  async update(id, title, maker, energy, protein, fat, carb) {
+  async update(userId, id, title, maker, energy, protein, fat, carb) {
     return await Product.update(
       { title, maker, energy, protein, fat, carb },
-      { where: { id } }
+      { where: { userId, id } }
     );
   }
 
-  async delete(id) {
-    const countIngredients = await Ingredient.count({
-      where: { productId: id },
+  async delete(userId, id) {
+    const product = await Product.findOne({
+      where: { userId, id },
+      include: Ingredient,
     });
-    if (countIngredients === 0) {
-      return await Product.destroy({ where: { id } });
+
+    if (product.toJSON().ingredients.length === 0) {
+      return await Product.destroy({ where: { userId, id } });
     }
     return await Product.update(
       { isDeleted: true },
       {
-        where: { id },
+        where: { userId, id },
       }
     );
   }
